+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # The `zksolc` changelog
 
+## [Unreleased]
+
+### Added
+
+- The EraVM disassembler
+
 ## [1.5.3] - 2024-08-27
 
 ### Added